+++ src/TrackRowInfoVis.js
@@ -21,7 +21,8 @@ const fieldTypeToVisComponent = {
  * @prop {number} top The top position of this view.
  * @prop {number} width The width of this view.
  * @prop {number} height The height of this view.
- * @prop {object[]} transformedRowInfo Array of JSON objects, one object for each row.
+ * @prop {object[]} rowInfo Array of JSON objects, one object for each sample, without filtering/sorting based on selected rows.
+ * @prop {object[]} transformedRowInfo The `rowInfo` array after transforming by filtering and sorting according to the selected rows.
  * @prop {object} fieldInfo The name and type of data field.
  * @prop {boolean} isLeft Is this view on the left side of the track?
  * @prop {object} rowSort The options for sorting rows.
@@ -38,6 +39,7 @@ export default function TrackRowInfoVis(props) {
         left, top, width, height,
         fieldInfo,
         isLeft,
+        rowInfo,
         transformedRowInfo,
         rowSort,
         rowFilter,
@@ -160,6 +162,7 @@ export default function TrackRowInfoVis(props) {
                     height,
                     isLeft,
                     fieldInfo,
+                    rowInfo,
                     transformedRowInfo,
                     titleSuffix,
                     onSortRows,