+++ CHANGELOG.md
@@ -11,10 +11,12 @@
 - Added a search suggestion autocomplete feature.
 - Added a feature to request for bind TFs in a certain interval using Cistrome DB API and render the result in a simple data table.
 - Added a context menu for filtering or highlighting rows based on a nomival value in categorical bar charts.
+- Added hover styles for the `TrackRowInfoVisLink` visualization component, so that when a link area is hovered, a gray background rect and a text underline are also rendered.
 
 ### Changed
 - Fixed bug which prevented wrapper options for sorting, filtering, highlighting from being used upon initial component render.
 - Added resizer elements to their containers and used `d3-drag` for the mouse events.
+- Compute color scale for the `TrackRowInfoVisNominalBar` categorical value visualization based on the full `rowInfo` rather than `transformedRowInfo` so that the original color mapping is kept after row filtering events.
 
 
 ## 0.1.0 - 02/19/20